core+gtk: streak helpers return u32, drop shell-side clamp
`Database::get_streak` and `get_best_streak` (plus their `_for_label`
variants and the `streak_filtered` / `best_streak_filtered`
internals) returned `Result<i64>`. The values are day counts —
always ≥ 0, and capped at u32::MAX (~11 million years). The
shell wrapper at `src/db/mod.rs:785-794` had to clamp via
`.max(0) as u32` on every call.

The i64 was a chrono artefact: nothing in the streak logic
called for signed arithmetic. Flipped:

  - `get_streak`, `get_streak_for_label`, `streak_filtered` →
    `Result<u32>`
  - `get_best_streak`, `get_best_streak_for_label`,
    `best_streak_filtered` → `Result<u32>`
  - Internal counters: `1i64` → `1u32`, `0` → `0u32`

Shell wrappers now pass the core return through unchanged
instead of clamping.

+++ meditate-core/src/db/sessions.rs
@@ -321,21 +321,21 @@ impl Database {
         Ok(())
     }
 
-    pub fn get_best_streak(&self) -> Result<i64> {
+    pub fn get_best_streak(&self) -> Result<u32> {
         self.best_streak_filtered(None)
     }
 
-    pub fn get_best_streak_for_label(&self, label_id: i64) -> Result<i64> {
+    pub fn get_best_streak_for_label(&self, label_id: i64) -> Result<u32> {
         self.best_streak_filtered(Some(label_id))
     }
 
-    fn best_streak_filtered(&self, label_filter: Option<i64>) -> Result<i64> {
+    fn best_streak_filtered(&self, label_filter: Option<i64>) -> Result<u32> {
         let days = self.distinct_session_days_ascending(label_filter)?;
         if days.is_empty() {
             return Ok(0);
         }
-        let mut best = 1i64;
-        let mut current = 1i64;
+        let mut best = 1u32;
+        let mut current = 1u32;
         for window in days.windows(2) {
             if window[1] == window[0].succ_opt().expect("date overflow") {
                 current += 1;
@@ -529,11 +529,11 @@ impl Database {
         Ok(days)
     }
 
-    pub fn get_streak(&self, today: chrono::NaiveDate) -> Result<i64> {
+    pub fn get_streak(&self, today: chrono::NaiveDate) -> Result<u32> {
         self.streak_filtered(today, None)
     }
 
-    pub fn get_streak_for_label(&self, today: chrono::NaiveDate, label_id: i64) -> Result<i64> {
+    pub fn get_streak_for_label(&self, today: chrono::NaiveDate, label_id: i64) -> Result<u32> {
         self.streak_filtered(today, Some(label_id))
     }
 
@@ -541,7 +541,7 @@ impl Database {
         &self,
         today: chrono::NaiveDate,
         label_filter: Option<i64>,
-    ) -> Result<i64> {
+    ) -> Result<u32> {
         let days = self.distinct_session_days_ascending(label_filter)?;
         let Some(&most_recent) = days.last() else {
             return Ok(0);
@@ -564,7 +564,7 @@ impl Database {
             return Ok(0);
         };
 
-        let mut count = 0;
+        let mut count = 0u32;
         for day in days.iter().rev() {
             if *day == expected {
                 count += 1;
